requirements.txt
aiofiles
aiohttp
python-dotenv
websockets